App functionality
There is now a confirm password box when creating a user. All fields are required that way the app doesnt crash when it tries to execute sql statements with empty parameters. Program checks for duplicate usernames in the database before adding a new user and rejects it if it is duplicate. Successful login page now displays all the information of the user when they login. There are links to navigate between pages on each page.