Search Bar
Removed one of the foreign keys from the youtubevideos table that was involving PostUser. There is now a primitive search bar on the user home page. You can search by a single first name or a first name and a last name of two different comedians seperated by a space. The user is redirected to another page that displays all the videos related to the search query in table format.