Search bar and favorite list
Added bootstrap to the search bar and the search results page so that it looks a lot more clean. There is now a button on the home page that brings the user to another page and shows them all the comedians on their favorite list. The user can delete comedians and add comedians to their favorite list. When adding a comedian, there is a drop down that only lets them select from comedians that are not already on their list. There is bootstrap on the homepage, search results page, and favorite list page. Added classes for comedians and favorite list objects.