Video Page Changes and Content Merge
Video page is now stylized and looking nice. There is an embedded video there so users can watch the videos on our website. The title, user, date, and video description are all displayed on the page in a organized fasion. There is a section to write a review if the user hasn't already done so. This section goes away if the user has already written one or its their own video. There is also an add comedian to favorites button. This button will dissapear if the user already has the comedian on their favorite list.